requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.23.5
opencv-python==4.7.0.72
mediapipe==0.9.1.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pdfplumber==0.10.3
python-docx==1.0.1
PyPDF2==3.0.1
groq==0.4.2
jinja2==3.1.2 